requirements.txt
streamlit==1.27.2
pandas==2.0.0
google-api-python-client==2.95.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0